backend/requirements.txt
# Install with the same interpreter you use to run: py -3.13 -m pip install -r requirements.txt
google-genai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
